refactor(tests): split 2 oversized files into 10 focused modules (phase8 slice)
test_validator_false_positives.py (1992 lines) → 6 files:
- test_validator_fp_wrapper_a/b/c.py — wrapper-node false-positive tests
- test_validator_fp_depth.py — depth-limit false-positive tests
- test_validator_fp_multifile.py — multi-file scenario tests
- test_validator_fp_boundary.py — boundary/edge-case tests

test_call_graph.py (1122 lines) → 4 files:
- test_call_graph_ast_a.py — FunctionRef, NodeText, WalkTree
- test_call_graph_ast_b.py — GetFuncName, ExtractCall, FindParentClass
- test_call_graph_integration.py — CallGraphBuild and integration tests
- test_call_graph_cached.py — CachedCallGraph, FileImpact, NodeTextUtf8

All files under 500 lines. 252 tests pass before and after.
Full unit suite (16 598 tests) green.
